Continue generic progress on the package lists.
The lists are still a mess. I will finish them this weekend. Once they
are finished, I am going to modify each package in the list to point to
a package in the AUR. Existing packages in the AUR are the current
priority. Once those packages are listed, work can begin on adding new
packages to the AUR.