Add packages from the official repositories.
This was a tough decision, but I have decided to create archtrack
versions of packages in the official repos. This will allow me to put
official packages in the archtrack groups. The archtrack versions of
these packages will be differentiated from the official versions with
'-archtrack' suffixes.

I might apply some security-oriented patches to these packages one day.