build: Build in the package directories. Make gpg-agent quit correctly.
The build directory was getting cluttered with build remnants. Now we
build where the PKGBUILD files are and move the packages to the build
directory afterwards.

gpg-agent was left running after the script quit. We now run the script
with gpg-agent from within the script itself. This makes gpg-agent exit
when the script does. It feels wrong but I couldn't find a better way.