[fix] cewl cd'ing to directory...
As said in #blackarch (freenode), if the PKGBUILD use a wrapper
that will cd to the application directory there is a high chance
it will write file to that same directory and not to the directory
that it was called from.

This fix ensures it.

Note: we might have to do it on others PKGBUILD too.