chipsec: Fix build issues during DKMS build stage
Makefile: Rewrite the variable for defining the actual location of
kernel from $(KERNEL_SRC_DIR) to $(KSRC) using a rather old but still
working method.

This is a rewrite compared to the [previous commit](https://github.com/BlackArch/blackarch/pull/3000)
and that it should be separate from the other PR; blackarch-keyring.

It will not compile. Apparently this is an issue that has already been
[upstreamed](https://github.com/chipsec/chipsec/issues/1086); however, it
will address the initial kernel source directory checking phase. Please
check comments for full disclosure.