chipsec: Updated fixes in light of previous commit
Makefile.patch: Fixed multiple issues with half-patched file that was in
my origin/work directory.
PKGBUILD: Bumped pkgrel, updated sha512sums, added changes as mentioned
below.

First and foremost, sorry for pushing a half-baked solution in previous
commit. I have addressed the Makefile.patch as well as updated the
PKGBUILD with new hashsums and changes as per suggestions in the
[commit](https://github.com/BlackArch/blackarch/commit/d3fe35ee50083368db528acb5f7c0850489bed5e).