Add -f, --force flag and fix duplicate entries
The script will now check if there already is a ba-*.desktop file
present. If the -f flag is not provided, the script will terminate.

If you want to force an overwrite, because e.g. you want to provide
an icon for the application use

```
-f, --force
```

this will update existing shasums in the PKGBUILD file

No duplicate entries should be made in the install section or at source

Signed-off-by: Stefan Venz <stefan.venz@protonmail.com>