lbd: Change gnu-netcat dependency to openbsd-netcat and adding missing dependency (#3644)
* Changed gnu-netcat dependency to openbsd-netcat

gnu-netcat does not provide any output when run as listener while openbsd-netcat does (i.e. Listening on port 4444...) and it makes the user aware that netcat is listening. gnu-netcat and openbsd-netcat cannot be installed together because in conflict but openbsd-netcat is better (used also on other pentesting distros).

* Added missing dnsutils dependency

lbd uses `host` command that is provided by `dnsutils` package in Arch environment. If `dnsutils` is not installed, the `host` command won't be found and `lbd` will return always the message `Checking for DNS-Loadbalancing:/usr/bin/lbd: line 31: host: command not found`.

* pkgrel++

* add ending newline

Co-authored-by: Alexandre ZANNI <16578570+noraj@users.noreply.github.com>