eresi: moved linux-headers dependency to optdepends (#3750)
* eresi: removed linux-headers dependency

According to the project page, there is no request about the usage of `linux-headers` package. The install of this package could be dangerous for those systems using different kernel types. it could break the system.

The build process goes smooth and no evidence of kernel modules usage during the building.

* eresi: removed linux-headers dependency

* Update PKGBUILD