pcileech: moved linux-headers dependency to optdepends (#3752)
* pcileech: removed linux-headers dependency

According to the project page, there is no request about the usage of `linux-headers` package. The install of this package could be dangerous for those systems using different kernel types. it could break the system.

The needed dependencies for pcileech are specified at https://github.com/ufrisk/pcileech/wiki/PCILeech-on-Linux

* pcileech: removed linux-headers dependency

* Update PKGBUILD